Add option to set exact backup times

+++ ceph/files/backup/ceph-backup-client-runner-call.sh
@@ -3,6 +3,19 @@
 # Script to call ceph-backup-runner.sh in for loop to backup all keyspaces.
 # This script is also able to rsync backed up data to remote host and perform clean up on historical backups
 
+SKIPCLEANUP=false
+while getopts "sf" opt; do
+  case $opt in
+    s)
+      echo "Cleanup will be skipped" >&2
+      SKIPCLEANUP=true
+      ;;
+    \?)
+      echo "Invalid option: -$OPTARG" >&2
+      ;;
+  esac
+done
+
 # Configuration
 # -------------
     BACKUPDIR="{{ backup.backup_dir }}/full"
@@ -13,16 +26,16 @@
 
     SCRIPTDIR="/usr/local/bin"
     KEEP={{ backup.client.full_backups_to_keep }}
-    HOURSFULLBACKUPLIFE={{ backup.client.hours_before_full }} # Lifetime of the latest full backup in seconds
-    RSYNCLOGDIR="/var/log/backups"
-    RSYNCLOG="/var/log/backups/ceph-rsync.log"
-
-
+    {%- if backup.client.backup_times is not defined %}
+    HOURSFULLBACKUPLIFE={{ backup.client.hours_before_full }} # Lifetime of the latest full backup in hours
     if [ $HOURSFULLBACKUPLIFE -gt 24 ]; then
         FULLBACKUPLIFE=$(( 24 * 60 * 60 ))
     else
         FULLBACKUPLIFE=$(( $HOURSFULLBACKUPLIFE * 60 * 60 ))
     fi
+    {%- endif %}
+    RSYNCLOGDIR="/var/log/backups"
+    RSYNCLOG="/var/log/backups/ceph-rsync.log"
 
 # Functions
 # ---------
@@ -75,6 +88,25 @@
 
 # Cleanup
 # ---------
+if [ $SKIPCLEANUP = false ] ; then
+    {%- if backup.client.backup_times is not defined %}
+    echo "----------------------------"
     echo "Cleanup. Keeping only $KEEP full backups"
     AGE=$(($FULLBACKUPLIFE * $KEEP / 60))
     find $BACKUPDIR -maxdepth 1 -type d -mmin +$AGE -execdir echo "removing: "$BACKUPDIR/{} \; -execdir rm -rf $BACKUPDIR/{} \;
+    {%- else %}
+    echo "----------------------------"
+    echo "Cleanup. Keeping only $KEEP full backups"
+    NUMBER_OF_FULL=`find $BACKUPDIR -maxdepth 1 -mindepth 1 -type d -print| wc -l`
+    FULL_TO_DELETE=$(( $NUMBER_OF_FULL - $KEEP ))
+    if [ $FULL_TO_DELETE -gt 0 ] ; then
+        cd $BACKUPDIR
+        ls -t | tail -n -$FULL_TO_DELETE | xargs -d '\n' rm -rf
+    else
+        echo "There are less full backups than required, not deleting anything."
+    fi
+    {%- endif %}
+else
+    echo "----------------------------"
+    echo "-s parameter passed. Cleanup was not triggered"
+fi
